src/requirements.txt
Flask
pymongo
marshmallow
python-dateutil